+++ spade_artifact/__init__.py
@@ -4,8 +4,9 @@
 
 __author__ = """Javi Palanca"""
 __email__ = "jpalanca@gmail.com"
-__version__ = "0.1.1"
-
+__version__ = "0.3.0"
 
 from .artifact import Artifact
 from .agent import ArtifactMixin
+
+__all__ = ["Artifact", "ArtifactMixin"]