+++ spade_artifact/artifact.py
@@ -1,26 +1,40 @@
 # -*- coding: utf-8 -*-
+import abc
 import asyncio
-import logging
-import sys
 import time
 from asyncio import Event
-from contextlib import suppress
-from typing import Union
+from typing import Union, Optional
 
-import aiosasl
-import aioxmpp
-from aiosasl import AuthenticationFailure
-from aioxmpp import ibr, XMPPCancelError, XMPPAuthError
-from aioxmpp.dispatcher import SimpleMessageDispatcher
+from slixmpp import JID
+from slixmpp.exceptions import IqError, _DEFAULT_ERROR_TYPES
+from slixmpp.stanza.message import Message as SlixmppMessage
 from loguru import logger
+from spade.agent import AuthenticationFailure, DisconnectedException
 from spade.container import Container
 from spade.message import Message
 from spade.presence import PresenceManager
+from spade.xmpp_client import XMPPClient
 from spade_pubsub import PubSubMixin
 
 
-class Artifact(PubSubMixin):
-    def __init__(self, jid, password, pubsub_server=None, verify_security=False):
+class AbstractArtifact(object, metaclass=abc.ABCMeta):
+    async def _hook_plugin_before_connection(self, *args, **kwargs):
+        """
+        Overload this method to hook a plugin before connection is done
+        """
+        pass
+
+    async def _hook_plugin_after_connection(self, *args, **kwargs):
+        """
+        Overload this method to hook a plugin after connection is done
+        """
+        pass
+
+
+class Artifact(PubSubMixin, AbstractArtifact):
+    def __init__(
+        self, jid, password, pubsub_server=None, port=5222, verify_security=False
+    ):
         """
         Creates an artifact
 
@@ -29,30 +43,30 @@ class Artifact(PubSubMixin):
           password (str): The password to connect to the server
           verify_security (bool): Wether to verify or not the SSL certificates
         """
-        self.jid = aioxmpp.JID.fromstr(jid)
+        self.jid = JID(jid)
         self.password = password
+        self.xmpp_port = port
         self.verify_security = verify_security
 
         self.pubsub_server = (
             pubsub_server if pubsub_server else f"pubsub.{self.jid.domain}"
         )
 
+        self.client: Optional[XMPPClient] = None
+        self.presence: Optional[PresenceManager] = None
+
         self._values = {}
 
-        self.conn_coro = None
-        self.stream = None
-        self.client = None
         self.message_dispatcher = None
-        self.presence = None
 
         self.container = Container()
         self.container.register(self)
+
         self.loop = self.container.loop
 
-        # self.loop = None #asyncio.new_event_loop()
-
-        self.queue = asyncio.Queue(loop=self.loop)
+        self.queue = asyncio.Queue()
         self._alive = Event()
+        self.subscriptions = {}
 
     def set_loop(self, loop):
         self.loop = loop
@@ -66,7 +80,16 @@ class Artifact(PubSubMixin):
         """
         self.container = container
 
-    def start(self, auto_register=True):
+    async def _hook_plugin_after_connection(self, *args, **kwargs):
+        try:
+            await super()._hook_plugin_after_connection(*args, **kwargs)
+        except AttributeError:
+            logger.debug("_hook_plugin_after_connection is undefined")
+
+        # Set the publication handler once the connection is established
+        self.pubsub.set_on_item_published(self.on_item_published)
+
+    async def start(self, auto_register: bool = True) -> None:
         """
         Tells the container to start this agent.
         It returns a coroutine or a future depending on whether it is called from a coroutine or a synchronous method.
@@ -74,7 +97,7 @@ class Artifact(PubSubMixin):
         Args:
             auto_register (bool): register the agent in the server (Default value = True)
         """
-        return self.container.start_agent(agent=self, auto_register=auto_register)
+        return await self._async_start(auto_register=auto_register)
 
     async def _async_start(self, auto_register=True):
         """
@@ -92,86 +115,90 @@ class Artifact(PubSubMixin):
 
         await self._hook_plugin_before_connection()
 
-        if auto_register:
-            await self._async_register()
-        self.client = aioxmpp.PresenceManagedClient(
-            self.jid,
-            aioxmpp.make_security_layer(
-                self.password, no_verify=not self.verify_security
-            ),
-            loop=self.loop,
-            logger=logging.getLogger(self.jid.localpart),
+        self.client = XMPPClient(
+            self.jid, self.password, self.verify_security, auto_register
         )
-
-        # obtain an instance of the service
-        self.message_dispatcher = self.client.summon(SimpleMessageDispatcher)
 
         # Presence service
-        self.presence = PresenceManager(self)
+        self.presence = PresenceManager(agent=self, approve_all=False)
 
         await self._async_connect()
-
-        # register a message callback here
-        self.message_dispatcher.register_callback(
-            aioxmpp.MessageType.CHAT,
-            None,
-            self._message_received,
-        )
-
         await self._hook_plugin_after_connection()
 
         # pubsub initialization
         try:
-            self._node = str(self.jid.bare())
+            self._node = str(self.jid.bare)
             await self.pubsub.create(self.pubsub_server, f"{self._node}")
-        except XMPPCancelError as e:
-            logger.info(f"Node {self._node} already registered")
-        except XMPPAuthError as e:
-            logger.error(f"Artifact {self._node} is not allowed to publish properties.")
+        except IqError as e:
+            if e.condition == _DEFAULT_ERROR_TYPES["conflict"]:
+                logger.info(f"Node {self._node} already registered")
+            elif e.condition == _DEFAULT_ERROR_TYPES["forbidden"]:
+                logger.error(
+                    f"Artifact {self._node} is not allowed to publish properties."
+                )
+            else:
+                logger.error(f"Error creating node: {e.format()}")
             raise e
 
         await self.setup()
         self._alive.set()
         asyncio.run_coroutine_threadsafe(self.run(), loop=self.loop)
 
-    async def _hook_plugin_before_connection(self, *args, **kwargs):
-        """
-        Overload this method to hook a plugin before connetion is done
-        """
-        try:
-            await super()._hook_plugin_before_connection(*args, **kwargs)
-        except AttributeError:
-            logger.debug("_hook_plugin_before_connection is undefined")
-
-    async def _hook_plugin_after_connection(self, *args, **kwargs):
-        """
-        Overload this method to hook a plugin after connetion is done
-        """
-        try:
-            await super()._hook_plugin_after_connection(*args, **kwargs)
-        except AttributeError:
-            logger.debug("_hook_plugin_after_connection is undefined")
-
     async def _async_connect(self):  # pragma: no cover
-        """ connect and authenticate to the XMPP server. Async mode. """
-        try:
-            self.conn_coro = self.client.connected()
-            aenter = type(self.conn_coro).__aenter__(self.conn_coro)
-            self.stream = await aenter
-            logger.info(f"Artifact {str(self.jid)} connected and authenticated.")
-        except aiosasl.AuthenticationFailure:
-            raise AuthenticationFailure(
-                "Could not authenticate the artifact. Check user and password or use auto_register=True"
+        """connect and authenticate to the XMPP server. Async mode."""
+
+        if self.client is not None:
+            self.client.connected_event = asyncio.Event()
+            self.client.disconnected_event = asyncio.Event()
+            self.client.failed_auth_event = asyncio.Event()
+
+            connected_task = asyncio.create_task(
+                self.client.connected_event.wait(), name="connected"
+            )
+            disconnected_task = asyncio.create_task(
+                self.client.disconnected_event.wait(), name="disconnected"
+            )
+            failed_auth_task = asyncio.create_task(
+                self.client.failed_auth_event.wait(), name="failed_auth"
             )
 
-    async def _async_register(self):  # pragma: no cover
-        """ Register the artifact in the XMPP server from a coroutine. """
-        metadata = aioxmpp.make_security_layer(None, no_verify=not self.verify_security)
-        query = ibr.Query(self.jid.localpart, self.password)
-        _, stream, features = await aioxmpp.node.connect_xmlstream(
-            self.jid, metadata, loop=self.loop
-        )
-        await ibr.register(stream, query)
+            self.client.add_event_handler(
+                "session_start", lambda _: self.client.connected_event.set()
+            )
+            self.client.add_event_handler(
+                "disconnected", lambda _: self.client.disconnected_event.set()
+            )
+            self.client.add_event_handler(
+                "failed_all_auth", lambda _: self.client.failed_auth_event.set()
+            )
+            self.client.add_event_handler("message", self._message_received)
+
+            _ = self.client.connect(host=self.jid.host, port=self.xmpp_port)
+
+            done, pending = await asyncio.wait(
+                [connected_task, disconnected_task, failed_auth_task],
+                return_when=asyncio.FIRST_COMPLETED,
+            )
+
+            for task in pending:
+                task.cancel()
+
+            for task in done:
+                await task
+
+                if task.get_name() == "failed_auth":
+                    raise AuthenticationFailure(
+                        "Could not authenticate the agent. Check user and password or use auto_register=True"
+                    )
+                elif task.get_name() == "disconnected":
+                    raise DisconnectedException(
+                        "Error during the connection with the server"
+                    )
+
+            logger.info(f"Agent {str(self.jid)} connected and authenticated.")
+
+        else:
+            raise RuntimeError("XMPPClient is not initialized.")
 
     async def setup(self):
         """
@@ -192,27 +219,23 @@ class Artifact(PubSubMixin):
 
     @property
     def name(self):
-        """ Returns the name of the artifact (the string before the '@') """
-        return self.jid.localpart
+        """Returns the name of the artifact (the string before the '@')"""
+        return self.jid.node
 
-    def stop(self):
+    async def stop(self) -> None:
         """
-        Stop the artifact
+        Stops this agent.
         """
         self.kill()
-        return self.loop.run_until_complete(self._async_stop())
+        return await self._async_stop()
 
     async def _async_stop(self):
-        """ Stops an artifact and kills all its behaviours. """
+        """Stops an artifact and kills all its behaviours."""
         if self.presence:
             self.presence.set_unavailable()
 
-        """ Discconnect from XMPP server. """
         if self.is_alive():
-            # Disconnect from XMPP server
-            self.client.stop()
-            aexit = self.conn_coro.__aexit__(*sys.exc_info())
-            await aexit
+            await self.client.disconnect()
             logger.info("Client disconnected.")
 
         self._alive.clear()
@@ -254,22 +277,19 @@ class Artifact(PubSubMixin):
         else:
             return None
 
-    def _message_received(self, msg):
+    def _message_received(self, msg) -> None:
         """
         Callback run when an XMPP Message is reveived.
-        The aioxmpp.Message is converted to spade.message.Message
+        The slixmpp.stanza.Message is converted to spade.message.Message
 
         Args:
-          msg (aioxmpp.Messagge): the message just received.
-
-        Returns:
-            asyncio.Future: a future of the append of the message.
+          msg (slixmpp.stanza.Messagge): the message just received.
 
         """
 
         msg = Message.from_node(msg)
         logger.debug(f"Got message: {msg}")
-        return asyncio.run_coroutine_threadsafe(self.queue.put(msg), self.loop)
+        self.queue.put_nowait(msg)
 
     async def send(self, msg: Message):
         """
@@ -281,8 +301,8 @@ class Artifact(PubSubMixin):
         if not msg.sender:
             msg.sender = str(self.jid)
             logger.debug(f"Adding artifact's jid as sender to message: {msg}")
-        aioxmpp_msg = msg.prepare()
-        await self.client.send(aioxmpp_msg)
+        slixmpp_msg = msg.prepare(self.client)
+        slixmpp_msg.send()
         msg.sent = True
 
     async def receive(self, timeout: float = None) -> Union[Message, None]:
@@ -321,7 +341,6 @@ class Artifact(PubSubMixin):
         return self.queue.qsize()
 
     def join(self, timeout=None):
-
         try:
             in_coroutine = asyncio.get_event_loop() == self.loop
         except RuntimeError:  # pragma: no cover
@@ -346,4 +365,44 @@ class Artifact(PubSubMixin):
                 raise TimeoutError
 
     async def publish(self, payload: str) -> None:
-        await self.pubsub.publish(self.pubsub_server, self._node, payload)
+        await self.pubsub.publish(
+            self.pubsub_server, self._node, payload, ifrom=self.jid.bare
+        )
+
+    def on_item_published(self, msg: SlixmppMessage):
+        """
+        Callback to handle an item published event.
+
+        Args:
+            jid (str): The JID of the publisher.
+            node (str): The node/topic from which the item was published.
+            item (object): The item that was published.
+            message (str, optional): Additional message or data associated with the publication.
+        """
+        node = msg["pubsub_event"]["items"]["node"]
+        if node in self.subscriptions:
+            item = msg["pubsub_event"]["items"]["item"]["payload"]
+            jid = msg["pubsub_event"]["items"]["item"]["publisher"]
+            self.subscriptions[node](jid, item)
+
+    async def link(self, target_artifact_jid, callback):
+        """
+        Subscribe to another artifact's publications.
+
+        Args:
+            target_artifact_jid (str): The JID of the target artifact to subscribe to.
+            callback (Callable): The callback to invoke when an item is published.
+        """
+        await self.pubsub.subscribe(self.pubsub_server, str(target_artifact_jid))
+        self.subscriptions[target_artifact_jid] = callback
+
+    async def unlink(self, target_artifact_jid):
+        """
+        Unsubscribe from another artifact's publications.
+
+        Args:
+            target_artifact_jid (str): The JID of the target artifact to unsubscribe from.
+        """
+        await self.pubsub.unsubscribe(self.pubsub_server, str(target_artifact_jid))
+        if target_artifact_jid in self.subscriptions:
+            del self.subscriptions[target_artifact_jid]