+++ tests/conftest.py
@@ -1,17 +1,16 @@
 import pytest
-from aioxmpp import JID
+from slixmpp.jid import JID
 
-from tests.factories import MockedConnectedArtifactAgentFactory
+from .factories import MockedConnectedArtifactAgentFactory
 
 
 @pytest.fixture
 def jid():
-    return JID.fromstr("friend@localhost/home")
+    return JID("friend@localhost/home")
 
 
 @pytest.fixture
-def agent():
+async def agent():
     agent = MockedConnectedArtifactAgentFactory()
-    future = agent.start()
-    future.result()
-    return agent
+    await agent.start()
+    yield agent