+++ tests/factories.py
@@ -1,8 +1,8 @@
 import factory
-from aioxmpp import PresenceShow, PresenceState
-from asynctest import CoroutineMock, Mock
-from spade.agent import Agent
+from unittest.mock import AsyncMock, Mock
+from spade.presence import PresenceShow
 
+from spade.agent import Agent
 from spade_artifact import Artifact, ArtifactMixin
 
 
@@ -13,10 +13,8 @@ class MockedConnectedArtifact(Artifact):
         super().__init__(*args, **kwargs)
         if status is None:
             status = {}
-        self._async_connect = CoroutineMock()
-        self._async_register = CoroutineMock()
-        self.conn_coro = Mock()
-        self.conn_coro.__aexit__ = CoroutineMock()
+        self._async_connect = AsyncMock()
+        self._async_register = AsyncMock()
 
         self.available = available
         self.show = show
@@ -26,13 +24,13 @@ class MockedConnectedArtifact(Artifact):
     async def _hook_plugin_after_connection(self, *args, **kwargs):
         await super()._hook_plugin_after_connection(*args, **kwargs)
         self.pubsub = Mock()
-        self.pubsub.create = CoroutineMock()
+        self.pubsub.create = AsyncMock()
 
     def mock_presence(self):
         show = self.show if self.show is not None else PresenceShow.NONE
-        available = self.available if self.available is not None else False
-        state = PresenceState(available, show)
-        self.presence.presenceserver.set_presence(state, self.status, self.priority)
+        status = self.status if self.status is not None else ""
+        priority = self.priority if self.priority is not None else 0
+        self.presence.set_presence(show=show, status=status, priority=priority)
 
     async def run(self):
         self.set("test_passed", True)
@@ -54,17 +52,14 @@ class MockedConnectedArtifactFactory(factory.Factory):
 class MockedConnectedArtifactAgent(ArtifactMixin, Agent):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._async_connect = CoroutineMock()
-        self._async_register = CoroutineMock()
-        self.conn_coro = Mock()
-        self.conn_coro.__aexit__ = CoroutineMock()
-        self.stream = Mock()
+        self._async_connect = AsyncMock()
+        self._async_register = AsyncMock()
 
     async def _hook_plugin_after_connection(self, *args, **kwargs):
         await super()._hook_plugin_after_connection(*args, **kwargs)
         self.pubsub = Mock()
-        self.pubsub.subscribe = CoroutineMock()
-        self.pubsub.unsubscribe = CoroutineMock()
+        self.pubsub.subscribe = AsyncMock()
+        self.pubsub.unsubscribe = AsyncMock()
 
 
 class MockedConnectedArtifactAgentFactory(factory.Factory):