+++ tests/test_artifact.py
@@ -2,21 +2,27 @@
 # -*- coding: utf-8 -*-
 
 """Tests for `spade_artifact` package."""
-from asynctest import Mock, CoroutineMock
+import asyncio
+from unittest.mock import Mock, AsyncMock, MagicMock
+
+import pytest
 from spade.message import Message
+from slixmpp import Message as SlixmppMessage
 
 from tests.factories import MockedConnectedArtifactFactory, MockedConnectedArtifact
 
 
-def test_run():
+async def test_run():
     artifact = MockedConnectedArtifactFactory()
-    future = artifact.start()
-    future.result()
-    artifact.join()
+    artifact.loop = asyncio.get_event_loop()
+    await artifact.start()
+
+    await artifact.join()
+
     assert artifact.get("test_passed")
 
 
-def test_setup():
+async def test_setup():
     class TestArtifact(MockedConnectedArtifact):
         def __init__(self, *args, **kwargs):
             super().__init__(*args, **kwargs)
@@ -29,28 +35,30 @@ def test_setup():
             self.kill()
 
     artifact = TestArtifact(jid="fake@jid", password="fake_password")
+    artifact.loop = asyncio.get_event_loop()
 
     assert artifact.value is False
-    future = artifact.start()
-    future.result()
-    artifact.join()
+    await artifact.start()
+    await artifact.join()
     assert artifact.value
 
 
-def test_name():
+async def test_name():
     artifact = MockedConnectedArtifactFactory()
     assert artifact.name == "fake"
 
 
-def test_is_alive():
+async def test_is_alive():
     artifact = MockedConnectedArtifactFactory()
+    artifact.loop = asyncio.get_event_loop()
+
     assert artifact.is_alive() is False
-    future = artifact.start()
-    future.result()
+    await artifact.start()
+
     assert artifact.is_alive()
 
 
-def test_set_get():
+async def test_set_get():
     artifact = MockedConnectedArtifactFactory()
 
     assert artifact.get("A_KEY") is None
@@ -64,61 +72,66 @@ def test_set_get():
     assert artifact.get("A_KEY") == 1234
 
 
-def test_send_msg():
+@pytest.mark.asyncio
+async def test_send_msg():
+    message = MagicMock()
+    message_prepare = MagicMock()
+    message.prepare.return_value = message_prepare
+    message_prepare.send = MagicMock()
+
     class A(MockedConnectedArtifact):
         async def run(self):
-            self.client = Mock()
-            self.client.send = CoroutineMock()
-            msg = Message()
-            await self.send(msg)
+            self.client = MagicMock()
+            await self.send(message)
             self.kill()
 
     artifact = A(jid="fakejid", password="fakesecret")
+    artifact.loop = asyncio.get_event_loop()
 
-    future = artifact.start()
-    future.result()
-    artifact.join()
+    await artifact.start()
+    await artifact.join()
 
-    assert artifact.client.send.called_with(Message().prepare())
+    assert message_prepare.send.called
+    message.prepare.assert_called_with(artifact.client)
 
 
-def test_receive():
+@pytest.mark.asyncio
+async def test_receive():
     class A(MockedConnectedArtifact):
         async def run(self):
             self.client = Mock()
-            self.client.send = CoroutineMock()
+            self.client.send = AsyncMock()
             self.msg = await self.receive(1)
             self.kill()
 
     artifact = A(jid="fakejid", password="fakesecret")
+    artifact.loop = asyncio.get_event_loop()
+    artifact._message_received(SlixmppMessage())
 
-    artifact._message_received(Message().prepare())
-
-    future = artifact.start()
-    future.result()
-    artifact.join()
+    await artifact.start()
+    await artifact.join()
 
     assert artifact.msg == Message()
 
 
-def test_mailbox_size():
+@pytest.mark.asyncio
+async def test_mailbox_size():
     class A(MockedConnectedArtifact):
         async def run(self):
             self.client = Mock()
-            self.client.send = CoroutineMock()
+            self.client.send = AsyncMock()
             self.msg = await self.receive(1)
             self.kill()
 
     artifact = A(jid="fakejid", password="fakesecret")
+    artifact.loop = asyncio.get_event_loop()
 
-    future = artifact._message_received(Message().prepare())
-    future.result()
+    artifact._message_received(SlixmppMessage())
 
     assert artifact.mailbox_size() == 1
 
-    future = artifact.start()
-    future.result()
-    artifact.join()
+    await artifact.start()
+    await artifact.join()
 
     assert artifact.msg == Message()